llm/requirements.txt
mistralai==1.7.0
dacite==1.9.2
fastapi==0.115.12
fastapi[standard]==0.115.12
pydantic==2.11.4
python-dotenv==1.1.0